[AS] removed static
