+++ src/test/java/com/adyen/mirakl/config/CucumberConfiguration.java
@@ -1,0 +1,16 @@
+package com.adyen.mirakl.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Configuration
+public class CucumberConfiguration {
+
+    @Bean
+    public Map<String, Object> cucumberMap() {
+        return new HashMap<>();
+    }
+}
