+++ src/test/java/com/adyen/mirakl/cucumber/CucumberTest.java
@@ -7,7 +7,7 @@ import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(plugin = "pretty", features = "src/test/features", tags = {"~@bug", "~@cucumber", "~@exclude"})
+@CucumberOptions(plugin = "pretty", features = "src/test/features", tags = {"~@bug", "@cucumber", "~@exclude"})
 public class CucumberTest  {
 
 }
