+++ src/test/java/com/adyen/mirakl/cucumber/stepdefs/helpers/hooks/CucumberHooks.java
@@ -1,23 +1,19 @@
 package com.adyen.mirakl.cucumber.stepdefs.helpers.hooks;
 
+import com.adyen.mirakl.cucumber.stepdefs.helpers.stepshelper.StepDefsHelper;
 import cucumber.api.java.Before;
 import org.awaitility.Awaitility;
 import org.awaitility.Duration;
 
-import java.util.HashMap;
-import java.util.Map;
-
-public class CucumberHooks {
-
-    public static Map<String, Object> cucumberMap;
-
-    @Before
-    public void initNewCucumberMap() {
-        cucumberMap = new HashMap<>();
-    }
+public class CucumberHooks extends StepDefsHelper {
 
     @Before
     public void setDefaultAwaitilityTimeOut() {
         Awaitility.setDefaultTimeout(Duration.FIVE_MINUTES);
+    }
+
+    @Before
+    public void clearCucumberMap() {
+        cucumberMap.clear();
     }
 }
